Tweak token vec pre-allocation heuristic
We increase the size of the initial token `Vec` by 1. This small change
has a significant impact on the performance for small expressions. For
the first 4 benchmarks there were gains of 26 to 40% in performance. For
larger sizes the variation is not very significant.

This is probably due to the fact that probably the initial allocation for
small expressions was not being large enough with the old heuristic formula,
causing the algorithm to reallocate the `Vec` during the process.